% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.dataset.R
\name{cal.dataset}
\alias{cal.dataset}
\title{Generate a synthetic dataset for clumped isotopes calibrations}
\usage{
cal.dataset(error = "S1", nobs = 1000)
}
\arguments{
\item{error}{Error scenario: low (S1), Intermediate (S2), or High (S3)}

\item{nobs}{Number of observations in the simulated dataset}
}
\description{
Generate a synthetic dataset for clumped isotopes calibrations
}
