% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.deming.R
\name{cal.deming}
\alias{cal.deming}
\title{Fit Deming regression models on a given calibration dataset}
\usage{
cal.deming(data, replicates, samples = NULL)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}

\item{samples}{Number of samples per bootstrap replicate}
}
\value{
a \code{data.frame} with replicate-level
        regression parameters
}
\description{
Fit Deming regression models on a given calibration dataset
}
