% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.york.R
\name{cal.york}
\alias{cal.york}
\title{Fit York regression models on a given calibration dataset}
\usage{
cal.york(data, replicates, samples = NULL)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}

\item{samples}{Number of samples per bootstrap replicate}
}
\value{
a \code{data.frame} with replicate-level
        regression parameters
}
\description{
Fit York regression models on a given calibration dataset
}
