% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_Bayesian.R
\name{fitClumpedRegressions}
\alias{fitClumpedRegressions}
\title{Bayesian regressions to calibrate the clumped isotopes paleothermometer using
\code{stan}.}
\usage{
fitClumpedRegressions(
  calibrationData,
  numSavedSteps = 3000,
  priors = "Informative"
)
}
\arguments{
\item{calibrationData}{The target calibration dataset.}

\item{numSavedSteps}{Number of MCMC iterations to save.}

\item{priors}{Either \code{Informative}, \code{Weak}, or
\code{Uninformative} on the slope and intercept.}

\item{samples}{Number of samples in the \code{calibrationData} to analyze.}
}
\description{
Bayesian regressions to calibrate the clumped isotopes paleothermometer using
\code{stan}.
}
