% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_Priors.R
\name{generatePriorDistCalibration}
\alias{generatePriorDistCalibration}
\title{Generate a dataset reflecting the priors used to run the analyses}
\usage{
generatePriorDistCalibration(prior, n = 1000)
}
\arguments{
\item{prior}{Informative or not}

\item{n}{number of observations to simulate}
}
\description{
Generate a dataset reflecting the priors used to run the analyses
}
