% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cal.prior.R
\name{plot.cal.prior}
\alias{plot.cal.prior}
\title{Generate a dataset reflecting the priors used to run the analyses}
\usage{
plot.cal.prior(prior, n = 1000)
}
\arguments{
\item{prior}{Informative or not}

\item{n}{number of observations to simulate}
}
\description{
Generate a dataset reflecting the priors used to run the analyses
}
