% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rec.priors.R
\name{plot.rec.priors}
\alias{plot.rec.priors}
\title{Generate a dataset reflecting the priors used to run the analyses}
\usage{
plot.rec.priors(prior, n = 1000)
}
\arguments{
\item{prior}{Informative or not}

\item{n}{number of observations to simulate}
}
\description{
Generate a dataset reflecting the priors used to run the analyses
}
