% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions_Frequentist.R
\name{predictTc}
\alias{predictTc}
\title{This function performs temp reconstruction (10^6/T^2 with T in K) for
multiple replicates of the same target.}
\usage{
predictTc(recData, obCal)
}
\arguments{
\item{recData}{Reconstruction dataset}

\item{obCal}{A \code{data.frame} summarizing the distribution of slopes and intercepts}
}
\description{
This function performs temp reconstruction (10^6/T^2 with T in K) for
multiple replicates of the same target.
}
