% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec.clumped.R
\name{rec.clumped}
\alias{rec.clumped}
\title{This function performs temp reconstruction (10^6/T^2 with T in K) for
multiple replicates of the same target.}
\usage{
rec.clumped(recData, obCal)
}
\arguments{
\item{recData}{Reconstruction dataset}

\item{obCal}{A \code{data.frame} summarizing the distribution of slopes and intercepts}
}
\value{
a \code{data.frame} with temperature reconstructions and the
        original values used in the reconstruction.
}
\description{
This function performs temp reconstruction (10^6/T^2 with T in K) for
multiple replicates of the same target.
}
