% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec.ion.bayesian.R
\name{rec.ion.bayesian}
\alias{rec.ion.bayesian}
\title{This function generates temperature predictions (in 10^6/T2) based on a
calibration dataset and target D47. It includes measured error in the target D47
and ion as an additional predictor.}
\usage{
rec.ion.bayesian(
  calModel,
  recData,
  iter = 1000,
  postcalsamples = NULL,
  MC = TRUE
)
}
\arguments{
\item{calModel}{The fitted Stan model from calibration.}

\item{recData}{The reconstruction dataset.}

\item{iter}{Number of sampling iterations.}

\item{postcalsamples}{Number of posterior calibration samples to use.}

\item{MC}{Enable multicore (default TRUE).}
}
\description{
This function generates temperature predictions (in 10^6/T2) based on a
calibration dataset and target D47. It includes measured error in the target D47
and ion as an additional predictor.
}
