% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_CI.R
\name{RegressionSingleCI}
\alias{RegressionSingleCI}
\title{This function is used to generate CI estimates at given intervals. It is currently
used for plotting in BayClump.}
\usage{
RegressionSingleCI(data, from, to, length.out = 100)
}
\arguments{
\item{data}{A \code{data.frame} with two columns named as beta and alpha.
This should be the result of bootstrapping or the posterior distribution
for a given calibration set.}

\item{from}{the lower limit in x.}

\item{to}{the upper limit in x.}

\item{length.out}{the number of breaks.}
}
\description{
This function is used to generate CI estimates at given intervals. It is currently
used for plotting in BayClump.
}
