% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_SimulateDS.R
\name{sim.clumped}
\alias{sim.clumped}
\title{Generate a synthetic dataset for clumped isotopes calibrations}
\usage{
sim.clumped(error = "S1", nobs = 1000)
}
\arguments{
\item{error}{Error scenario: low (S1), Intermediate (S2), or High (S3)}

\item{nobs}{Number of observations in the simulated dataset}
}
\description{
Generate a synthetic dataset for clumped isotopes calibrations
}
