% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_Deming.R
\name{simulateDeming}
\alias{simulateDeming}
\title{Fit Deming regression models on a given calibration dataset}
\usage{
simulateDeming(data, replicates)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}
}
\description{
Fit Deming regression models on a given calibration dataset
}
