% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_WeightedOLS.R
\name{simulateLM_inverseweights}
\alias{simulateLM_inverseweights}
\title{Fit weighted OLS regression models on a given calibration dataset}
\usage{
simulateLM_inverseweights(data, replicates)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}
}
\description{
Fit weighted OLS regression models on a given calibration dataset
}
