% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_OLS.R
\name{simulateLM_measured}
\alias{simulateLM_measured}
\title{Fit OLS regression models on a given calibration dataset}
\usage{
simulateLM_measured(data, replicates)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}
}
\description{
Fit OLS regression models on a given calibration dataset
}
