% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_York.R
\name{simulateYork_measured}
\alias{simulateYork_measured}
\title{Fit York regression models on a given calibration dataset}
\usage{
simulateYork_measured(data, replicates)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}
}
\description{
Fit York regression models on a given calibration dataset
}
