% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.york.R
\name{cal.york}
\alias{cal.york}
\title{Fit York regression models on a given calibration dataset}
\usage{
cal.york(data, replicates)
}
\arguments{
\item{data}{The calibration dataset}

\item{replicates}{Number of bootstrap replicates}
}
\description{
Fit York regression models on a given calibration dataset
}
